Huge refactor, lots of churn. nothing compiles
