+++ src/main/java/org/openstreetmap/atlas/geography/atlas/change/merge/FeatureChangeMergeGroup.java
@@ -1,5 +1,8 @@
-package org.openstreetmap.atlas.geography.atlas.change;
+package org.openstreetmap.atlas.geography.atlas.change.merge;
 
+import org.openstreetmap.atlas.geography.atlas.change.Change;
+import org.openstreetmap.atlas.geography.atlas.change.ChangeType;
+import org.openstreetmap.atlas.geography.atlas.change.FeatureChange;
 import org.openstreetmap.atlas.geography.atlas.items.ItemType;
 
 /**
